Add BasicControls and reorganize sample code
Added BasicControls.java for robot drive control. Deleted HelloFather.java. Moved GamepadPractice, HelloWorld, and VariablePractice to sampleCode package for better organization.
