--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/HelloFather.java
@@ -1,18 +1,0 @@
-package org.firstinspires.ftc.teamcode;
-
-import com.qualcomm.robotcore.eventloop.opmode.OpMode;
-import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
-
-@TeleOp
-public class HelloFather extends OpMode {
-
-    @Override
-    public void init() {
-        telemetry.addData("Hello", "Mr.", "Name");
-    }
-
-    @Override
-    public void loop() {
-        
-    }
-}
